+++ src/EVM/Types.hs
@@ -479,17 +479,30 @@ instance Ord Prop where
   PBool a <= PBool b = a <= b
   PEq (a :: Expr x) (b :: Expr x) <= PEq (c :: Expr y) (d :: Expr y)
     = case eqT @x @y of
-       Just Refl -> a <= c && b <= d
-       Nothing -> False
-  PLT a b <= PLT c d = a <= c && b <= d
-  PGT a b <= PGT c d = a <= c && b <= d
-  PGEq a b <= PGEq c d = a <= c && b <= d
-  PLEq a b <= PLEq c d = a <= c && b <= d
+       Just Refl -> a <= c || b <= d
+       Nothing -> toNum a <= toNum c
+  PLT a b <= PLT c d = a <= c || b <= d
+  PGT a b <= PGT c d = a <= c || b <= d
+  PGEq a b <= PGEq c d = a <= c || b <= d
+  PLEq a b <= PLEq c d = a <= c || b <= d
   PNeg a <= PNeg b = a <= b
-  PAnd a b <= PAnd c d = a <= c && b <= d
-  POr a b <= POr c d = a <= c && b <= d
-  PImpl a b <= PImpl c d = a <= c && b <= d
-  _ <= _ = False
+  PAnd a b <= PAnd c d = a <= c || b <= d
+  POr a b <= POr c d = a <= c || b <= d
+  PImpl a b <= PImpl c d = a <= c || b <= d
+  a <= b = asNum a <= asNum b
+    where
+      asNum :: Prop -> Int
+      asNum (PBool {}) = 0
+      asNum (PEq   {}) = 1
+      asNum (PLT   {}) = 2
+      asNum (PGT   {}) = 3
+      asNum (PGEq  {}) = 4
+      asNum (PLEq  {}) = 5
+      asNum (PNeg  {}) = 6
+      asNum (PAnd  {}) = 7
+      asNum (POr   {}) = 8
+      asNum (PImpl {}) = 9
+
 
 
 isPBool :: Prop -> Bool